12/12 commodity part: + application.prop: hibernate.enable_lazy_load, gamil account + features for shopping cart, orders, discounts

+++ FinalTopic/src/main/webapp/WEB-INF/jsp/commodity/commodityList.jsp
@@ -6,15 +6,11 @@
 <head>
 <meta charset="UTF-8">
 <title>Commodity List</title>
-<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-
+<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
+<link rel="stylesheet" href="//code.jquery.com/ui/1.13.0/themes/base/jquery-ui.css">
 <style>
-body{
-	background-color: #f2f2f2;;
-	padding: 10px;
-}
 #showComm{
-	font-size: 15px;
+	font-size: 12px;
 }
 
 th{
@@ -23,58 +19,71 @@ th{
 </style>
 </head>
 <body>
-<div><h4>商品管理</h4><a href='commodityForm?id=1'>新增商品</a></div>
-<c:if test="${not empty successMsg }">
-<div class="alert alert-warning alert-dismissible fade show" role="alert">
-  <strong>新增/修改商品成功!</strong>
-  <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
-</div>
-</c:if>
-<br>
-<div id='mySearch' class="row row-cols-2 row-cols-lg-5 g-2 g-lg-3">
-	<div class="col form-floating">
-		<input type='text' id='itemId' class='searchinput form-control'>
-		<label>商品編號</label>
-	</div>
-	<div class="col form-floating">
-		<input type='text' id='itemName' class='searchinput form-control'>
-		<label>商品名稱</label>
-	</div>
-	<div class="col form-floating">
-		<select id='itemType' class='searchinput form-select'></select>
-		<label>商品類別</label>
-	</div>
-	<div class="col form-floating">
-		<select id='onSale' class='searchinput form-select'><option label='請選擇' value='' /><option label='是' value='true'/><option label='否' value='false'/></select>
-		<label>是否上架</label>
-	</div>
-	<div class="col form-floating">
-		<input type='date' id='lastUpdated' class='searchinput form-control'>
-		<label>建立日期</label>
-	</div>
-	<div class="col form-floating">
-		<input type='text' id='vendorName' class='searchinput form-control'>
-		<label>供應商</label>
-	</div>
-	<div class="col">
-		<button class='btn btn-outline-primary' onclick='search()'>套用搜尋</button>
-		<button class="btn btn-outline-secondary" onclick='clearout()'>清除篩選</button>
-	</div>
-</div>
-<hr>
-<table id="showpage">
-   <tr>
-      <td>共<strong id='eleTTL'></strong>筆資料, 頁數<strong id='pgTTL'></strong></td>
-      <td id='pages' colspan="8" align="left"></td>
-   </tr>
-</table>
-<div class='table-responsive'>
-	<table id='showComm' border='1' style="table-layout:fixed;" class="table table-hover">
-	<thead class="table-light"><tr><th style='width:5%'>#</th><th>商品名稱</th><th style='width:5%'>庫存數量</th><th style='width:5%'>單價</th><th style='width:5%'>商品類別</th><th style='width:35%'>商品描述</th><th style='width:8%'>建立日期</th><th style='width:5%'>上架中</th><th style='width:10%'>供應商</th><th>操作</th></tr></thead>
-	<tbody id='tb'></tbody>
-	</table>
-</div>
 
+<div class="wrapper">
+<jsp:include page="/WEB-INF/jsp/commodity/00_sidebar.jsp" /> 
+<div class="main-panel">
+<c:import url="/adminnavbar"></c:import> 
+<div class="content">
+
+<div class="container-fluid">
+	<div>
+		<h4>商品管理</h4><a class="btn btn-primary" href='commodityForm?id=1'>新增商品</a>
+	</div>
+	
+	<c:if test="${not empty successMsg }">
+		<div class="alert alert-success alert-dismissible fade show" role="alert">
+  		<strong>${successMsg}</strong>
+  		<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+		</div>
+	</c:if>
+	
+	<br>
+	<div id='mySearch' class="row row-cols-2 row-cols-lg-5 g-2 g-lg-3">
+		<div class="col form-floating">
+			<input type='text' id='itemId' class='searchinput form-control'>
+			<label>商品編號</label>
+		</div>
+		<div class="col form-floating">
+			<input type='text' id='itemName' class='searchinput form-control'>
+			<label>商品名稱</label>
+		</div>
+		<div class="col form-floating">
+			<select id='itemType' class='searchinput form-select'></select>
+			<label>商品類別</label>
+		</div>
+		<div class="col form-floating">
+			<select id='onSale' class='searchinput form-select'><option label='請選擇' value='' /><option label='是' value='true'/><option label='否' value='false'/></select>
+			<label>是否上架</label>
+		</div>
+		<div class="col form-floating">
+			<input type='date' id='lastUpdated' class='searchinput form-control'>
+			<label>建立日期</label>
+		</div>
+		<div class="col form-floating ui-widget">
+			<input type='text' id='vendorName' class='searchinput form-control'>
+			<label for='vendorName'>供應商</label>
+		</div>
+		<div class="col">
+			<button class='btn btn-primary mt-1' onclick='search()'>套用搜尋</button>
+			<button class="btn btn-secondary mt-1" onclick='clearout()'>清除篩選</button>
+		</div>
+	</div>
+	<hr>
+	<div class='row align-middle'>
+	   <p>共<strong id='eleTTL'></strong>筆資料, 頁數<strong id='pgTTL'></strong><span id='pages' class='float-end'></span></p>
+	</div>
+	<div class='table-responsive'>
+		<table id='showComm' border='1' style="table-layout:fixed;" class="table table-hover">
+		<thead class="table-light"><tr><th style='width:5%'>#</th><th>商品名稱</th><th style='width:5%'>庫存數量</th><th style='width:5%'>單價</th><th style='width:5%'>商品類別</th><th style='width:35%'>商品描述</th><th style='width:8%'>建立日期</th><th style='width:5%'>供應商</th><th style='width:10%'>狀態</th><th>操作</th></tr></thead>
+		<tbody id='tb'></tbody>
+		</table>
+	</div>
+	
+</div>
+</div>
+</div>
+</div>
 <script src="http://code.jquery.com/jquery-1.12.4.min.js"></script>
 <script src="https://code.jquery.com/ui/1.13.0/jquery-ui.js"></script>
 <script type="text/javascript">
@@ -128,6 +137,9 @@ th{
 	}
 	
 	function del(id){
+		if(!window.confirm('是否刪除商品 #'+ id)){
+			return;
+		}
 		$.ajax({
 			type:'delete',
 			url: 'commodity/'+id,
@@ -161,12 +173,11 @@ th{
 			};
 		$.ajax({
 			type: 'post',
-			url: 'commoditytest/'+currentPage,
+			url: 'commodity/'+currentPage,
 			contentType: 'application/json',
 			dataTyp: 'json',
 			data: JSON.stringify(data),
 			success: function (data){
-				
 				let eleTTL = data.pageInfo[1];
 				var tBody = $('#tb');
 				tBody.empty("");
@@ -191,8 +202,15 @@ th{
 				    let eleTb = document.getElementById("tb");
 					$.each(arr, function(idx, value){
 						let eleTr = document.createElement("tr");
-						let str = "<td>"+ value.itemId+ "</td><td>"+ value.itemName +"</td><td>"+ value.itemStock +"</td><td>"+ value.unitPrice +"</td><td>"+ value.itemType.typeName +"</td><td style='word-wrap:break-word;'>"+ value.itemDesc +"</td><td>"+ value.lastUpdated + "</td><td>"+ value.onSale + "</td><td>"+ value.vendorName +"</td><td>";
-						str += "<button class='btn btn-primary btn-sm' onclick=\"location.href='commodityForm?id="+ value.itemId +"';\">編輯</button><button class='btn btn-danger btn-sm' onclick='del("+ value.itemId +")'>刪除</button>";
+						let str = "<td>"+ value.itemId+ "</td><td>"+ value.itemName +"</td><td>"+ value.itemStock +"</td><td>"+ value.unitPrice +"</td><td>"+ 
+						value.itemType.typeName +"</td><td style='word-wrap:break-word;'>"+ value.itemDesc +"</td><td>"+ value.lastUpdated + "</td><td>"+ value.vendorName +"</td>";
+						if (value.onSale == true){
+							str += "<td><span class='badge rounded-pill bg-success'>上架中</span></td>";
+						}else{
+							str += "<td><span class='badge rounded-pill bg-secondary'>停售</span></td>";
+						}
+						
+						str += "<td><button class='btn btn-primary btn-sm mx-1' onclick=\"location.href='commodityForm?id="+ value.itemId +"';\">編輯</button><button class='btn btn-danger btn-sm mx-1' onclick='del("+ value.itemId +")'>刪除</button>";
 						eleTr.innerHTML=str;
 						trFrag.appendChild(eleTr);
 						eleTb.appendChild(trFrag);
@@ -206,24 +224,6 @@ th{
 		$('#mySearch').find('.searchinput').val('');
 		currentPage = 1;
 		search();
-	}
-	
-	function addToCart(id){
-		$.ajax({
-			url: "addToCart",
-			type: "POST",
-			data: {id: id,
-					qty: 1},
-			success: function(data){
-						alert(data);
-						let arr = Object.values(data);
-						let volume = 0;
-						$.each(arr, function(idx, value){
-							volume += value.orderQty;
-						})
-						$("#cart").html(volume);
-			}	
-		})
 	}
 	
 	$(".alert>button").on("click", function(){
